refactor(cloud): isolate store database boundary
